requirements.txt
python-telegram-bot==21.5
anthropic==0.34.0
aiohttp==3.10.5
python-docx==1.1.2

# PDF reading
pdfplumber==0.11.4
pdf2image==1.17.0

# OCR (scanned docs)
pytesseract==0.3.13
Pillow==10.4.0

# Office formats
openpyxl==3.1.5
docx2txt==0.8

# Env
python-dotenv==1.0.1
